vhost: fix compilation issue when vhost debug enabled
This patch fixes below compilation issue when
RTE_LIBRTE_VHOST_DEBUG is enabled:

In file included from ./lib/librte_vhost/virtio_net.c:50:0:
./lib/librte_vhost/virtio_net.c: In function ‘copy_mbuf_to_desc’:
./lib/librte_vhost/vhost.h:205:42: warning: format ‘%d’ expects argument of type ‘int’, but argument 5 has type ‘uint64_t {aka long unsigned int}’ [-Wformat=]
   snprintf(packet, VHOST_MAX_PRINT_BUFF, "(%d) Header size %d: ", (device->vid), (size)); \
                                          ^
./lib/librte_vhost/virtio_net.c:301:4: note: in expansion of macro ‘PRINT_PACKET’
    PRINT_PACKET(dev, (uintptr_t)dst, len, 0);
    ^~~~~~~~~~~~

Fixes: 266a4dae840e ("vhost: handle virtually non-contiguous buffers in Tx")
Fixes: a2bc90bb969a ("vhost: handle virtually non-contiguous buffers in Rx")
Fixes: 1e92b3cec1d7 ("vhost: handle virtually non-contiguous buffers in Rx-mrg")

Acked-by: Luca Boccassi <bluca@debian.org>

+++ lib/librte_vhost/virtio_net.c
@@ -297,7 +297,7 @@ copy_mbuf_to_desc(struct virtio_net *dev, struct vhost_virtqueue *vq,
 			rte_memcpy((void *)(uintptr_t)dst,
 					(void *)(uintptr_t)src, len);
 
-			PRINT_PACKET(dev, (uintptr_t)dst, len, 0);
+			PRINT_PACKET(dev, (uintptr_t)dst, (uint32_t)len, 0);
 			vhost_log_write(dev, guest_addr, len);
 			remain -= len;
 			guest_addr += len;
@@ -796,7 +796,7 @@ copy_mbuf_to_desc_mergeable(struct virtio_net *dev, struct vhost_virtqueue *vq,
 							len);
 
 					PRINT_PACKET(dev, (uintptr_t)dst,
-							len, 0);
+							(uint32_t)len, 0);
 					vhost_log_write(dev, guest_addr, len);
 
 					remain -= len;
@@ -1234,7 +1234,7 @@ copy_desc_to_mbuf(struct virtio_net *dev, struct vhost_virtqueue *vq,
 	rte_prefetch0((void *)(uintptr_t)(desc_addr + desc_offset));
 
 	PRINT_PACKET(dev, (uintptr_t)(desc_addr + desc_offset),
-			desc_chunck_len, 0);
+			(uint32_t)desc_chunck_len, 0);
 
 	mbuf_offset = 0;
 	mbuf_avail  = m->buf_len - RTE_PKTMBUF_HEADROOM;
@@ -1322,7 +1322,7 @@ copy_desc_to_mbuf(struct virtio_net *dev, struct vhost_virtqueue *vq,
 			desc_avail  = desc->len;
 
 			PRINT_PACKET(dev, (uintptr_t)desc_addr,
-					desc_chunck_len, 0);
+					(uint32_t)desc_chunck_len, 0);
 		} else if (unlikely(desc_chunck_len == 0)) {
 			desc_chunck_len = desc_avail;
 			desc_gaddr += desc_offset;
@@ -1337,7 +1337,7 @@ copy_desc_to_mbuf(struct virtio_net *dev, struct vhost_virtqueue *vq,
 			desc_offset = 0;
 
 			PRINT_PACKET(dev, (uintptr_t)desc_addr,
-					desc_chunck_len, 0);
+					(uint32_t)desc_chunck_len, 0);
 		}
 
 		/*
